fix useDebugFlags, use SelectAIModelMenu
