+++ quadratic-client/src/app/ai/components/SelectAIModelMenu.tsx
@@ -1,12 +1,17 @@
 import { useAIModel } from '@/app/ai/hooks/useAIModel';
 import { useDebugFlags } from '@/app/debugFlags/useDebugFlags';
-import { LightbulbIcon } from '@/shared/components/Icons';
+import { AIIcon, ArrowDropDownIcon, LightbulbIcon } from '@/shared/components/Icons';
+import { ROUTES } from '@/shared/constants/routes';
+import { Button } from '@/shared/shadcn/ui/button';
 import {
   DropdownMenu,
   DropdownMenuCheckboxItem,
   DropdownMenuContent,
   DropdownMenuTrigger,
 } from '@/shared/shadcn/ui/dropdown-menu';
+import { Label } from '@/shared/shadcn/ui/label';
+import { Popover, PopoverContent, PopoverTrigger } from '@/shared/shadcn/ui/popover';
+import { RadioGroup, RadioGroupItem } from '@/shared/shadcn/ui/radio-group';
 import { Toggle } from '@/shared/shadcn/ui/toggle';
 import { TooltipPopover } from '@/shared/shadcn/ui/tooltip';
 import { cn } from '@/shared/shadcn/utils';
@@ -14,27 +19,35 @@ import { CaretDownIcon } from '@radix-ui/react-icons';
 import mixpanel from 'mixpanel-browser';
 import { MODELS_CONFIGURATION } from 'quadratic-shared/ai/models/AI_MODELS';
 import type { AIModelConfig, AIModelKey } from 'quadratic-shared/typesAndSchemasAI';
-import { memo, useCallback, useMemo } from 'react';
+import { memo, useCallback, useMemo, useState } from 'react';
+import { Link } from 'react-router';
 
 interface SelectAIModelMenuProps {
   loading: boolean;
   textareaRef: React.RefObject<HTMLTextAreaElement | null>;
 }
 
+const modelsById = {
+  basic: { label: 'Basic', description: 'good for everyday tasks' },
+  pro: { label: 'Pro', description: 'smartest and most capable' },
+};
+
 export const SelectAIModelMenu = memo(({ loading, textareaRef }: SelectAIModelMenuProps) => {
+  const debugFlags = useDebugFlags();
+  const debug = useMemo(() => debugFlags.getFlag('debug'), [debugFlags]);
+
   const [selectedModel, setSelectedModel, selectedModelConfig, thinkingToggle, setThinkingToggle] = useAIModel();
-  const { getFlag } = useDebugFlags();
   const modelConfigs = useMemo(() => {
     const configs = Object.entries(MODELS_CONFIGURATION) as [AIModelKey, AIModelConfig][];
 
     // enable all models in debug mode
-    if (getFlag('debug')) {
+    if (debug) {
       return configs;
     }
 
     // only show enabled models in production
     return configs.filter(([_, config]) => config.enabled);
-  }, [getFlag]);
+  }, [debug]);
 
   const canToggleThinking = useMemo(
     () => selectedModelConfig.thinkingToggle !== undefined,
@@ -61,6 +74,10 @@ export const SelectAIModelMenu = memo(({ loading, textareaRef }: SelectAIModelMe
     [modelConfigs, selectedModel, setSelectedModel, setThinkingToggle]
   );
 
+  const isFreeUser = false; // TODO: Pull this from...? will probably have to update API
+  const [activeModel, setActiveModel] = useState<keyof typeof modelsById>(isFreeUser ? 'basic' : 'pro');
+  const activeModelLabel = modelsById[activeModel].label;
+
   return (
     <>
       {canToggleThinking && (
@@ -82,7 +99,7 @@ export const SelectAIModelMenu = memo(({ loading, textareaRef }: SelectAIModelMe
         </TooltipPopover>
       )}
 
-      {getFlag('debug') && (
+      {debug && (
         <DropdownMenu>
           <DropdownMenuTrigger
             disabled={loading}
@@ -122,8 +139,7 @@ export const SelectAIModelMenu = memo(({ loading, textareaRef }: SelectAIModelMe
                   >
                     <div className="flex w-full items-center justify-between text-xs">
                       <span className="pr-4">
-                        {(getFlag('debug') ? `${modelConfig.enabled ? '' : '(debug) '}${provider} - ` : '') +
-                          displayName}
+                        {(debug ? `${modelConfig.enabled ? '' : '(debug) '}${provider} - ` : '') + displayName}
                       </span>
                     </div>
                   </DropdownMenuCheckboxItem>
@@ -132,6 +148,52 @@ export const SelectAIModelMenu = memo(({ loading, textareaRef }: SelectAIModelMe
           </DropdownMenuContent>
         </DropdownMenu>
       )}
+
+      <Popover>
+        {/* Needs a min-width or it shifts as the popover closes */}
+        <PopoverTrigger className="group flex min-w-24 items-center justify-end gap-0 text-right">
+          Model: {activeModelLabel} <ArrowDropDownIcon className="group-[[aria-expanded=true]]:rotate-180" />
+        </PopoverTrigger>
+
+        <PopoverContent className="flex w-80 flex-col gap-2">
+          <div className="mt-2 flex flex-col items-center">
+            <AIIcon className="mb-2 text-primary" size="lg" />
+
+            <h4 className="text-lg font-semibold">AI models</h4>
+
+            <p className="text-sm text-muted-foreground">Choose the best fit for your needs.</p>
+          </div>
+
+          <form className="flex flex-col gap-1 rounded border border-border text-sm">
+            <RadioGroup
+              value={activeModel}
+              onValueChange={(val) => setActiveModel(val as keyof typeof modelsById)}
+              className="flex flex-col gap-0"
+            >
+              {Object.entries(modelsById).map(([id, { label, description }], i) => (
+                <Label
+                  className={cn(
+                    'cursor-pointer px-4 py-3 has-[:disabled]:cursor-not-allowed has-[[aria-checked=true]]:bg-accent has-[:disabled]:text-muted-foreground',
+                    i !== 0 && 'border-t border-border'
+                  )}
+                  key={id}
+                >
+                  <strong className="font-bold">{label}</strong>: <span className="font-normal">{description}</span>
+                  <RadioGroupItem value={id} className="float-right ml-auto" disabled={isFreeUser} />
+                </Label>
+              ))}
+            </RadioGroup>
+          </form>
+
+          {isFreeUser && (
+            <Button variant="link" asChild>
+              <Link to={ROUTES.ACTIVE_TEAM_SETTINGS} target="_blank">
+                Upgrade to Pro
+              </Link>
+            </Button>
+          )}
+        </PopoverContent>
+      </Popover>
     </>
   );
 });
