+++ quadratic-client/src/shared/hooks/useRemoveInitialLoadingUI.ts
@@ -1,9 +1,10 @@
 import { useDebugFlags } from '@/app/debugFlags/useDebugFlags';
 import mixpanel from 'mixpanel-browser';
-import { useLayoutEffect } from 'react';
+import { useLayoutEffect, useMemo } from 'react';
 
 export function useRemoveInitialLoadingUI() {
-  const { getFlag } = useDebugFlags();
+  const debugFlags = useDebugFlags();
+  const debug = useMemo(() => debugFlags.getFlag('debug'), [debugFlags]);
 
   useLayoutEffect(() => {
     // Get the initial start time (in ms) so we can track the load time
@@ -18,7 +19,7 @@ export function useRemoveInitialLoadingUI() {
     }
     const startTimeMs = Number(startTime);
     const loadTimeMs = Date.now() - startTimeMs;
-    if (getFlag('debug')) {
+    if (debug) {
       console.log(`Loading time: ${loadTimeMs}ms`);
     }
     const route = window.location.pathname + window.location.search;
@@ -26,6 +27,6 @@ export function useRemoveInitialLoadingUI() {
       route,
       loadTimeMs,
     });
-  }, [getFlag]);
+  }, [debug]);
   return null;
 }
